Added functionality to the edit button
When the edit button is selected, the edit and delete buttons for that particular task will now disappear. A textbox will appear containing the value of the todo task and a rename button now appear. Upon clicking the rename button, the task's name is now "edited" and stays in the same position.